Refactor CircularSlider to its own file
